Staging

+++ database/migrations/2014_10_12_000000_create_users_table.php
@@ -11,23 +11,25 @@ return new class extends Migration
      */
     public function up(): void
     {
-        Schema::create('users', function (Blueprint $table) {
-            $table->id();
-            $table->string('first_name');
-            $table->string('last_name');
-            $table->string('email')->unique();
-            $table->timestamp('email_verified_at')->nullable();
-            $table->string('password');
-            $table->boolean('active')->default(0);
-            $table->timestamp('activated_at')->nullable();
-            $table->boolean('verified')->default(0);
-            $table->timestamp('verified_at')->nullable();
-            $table->timestamp('gender')->nullable();
-            $table->foreignId('zone_id')->references('id')->on('zones')->onDelete('cascade');
-            $table->rememberToken();
-            $table->timestamps();
-            $table->softDeletes();
-        });
+        if (!Schema::hasTable('users')) {
+            Schema::create('users', function (Blueprint $table) {
+                $table->id();
+                $table->string('first_name');
+                $table->string('last_name');
+                $table->string('email')->unique();
+                $table->timestamp('email_verified_at')->nullable();
+                $table->string('password');
+                $table->boolean('active')->default(0);
+                $table->timestamp('activated_at')->nullable();
+                $table->boolean('verified')->default(0);
+                $table->timestamp('verified_at')->nullable();
+                $table->enum('gender', ['male', 'female'])->nullable();
+                $table->foreignId('zone_id')->references('id')->on('zones')->onDelete('cascade');
+                $table->rememberToken();
+                $table->timestamps();
+                $table->softDeletes();
+            });
+        }
     }
 
     /**
